drop support for beta functionality
The Pro client doesn't support any beta services at the moment.
Additionally, we want to update our beta service logic for the next beta
service we support. Due to that, we are removing the current code for
beta support. However, we are keeping the beta related CLI entrypoints
for compatibility reasons.