tests: relax regex to check for installed esm versions
On some behave tests, we check that the esm version of a package is
installed, whether because we explicitly ask for it, or we expect it to
happen automatically.

However, the esm version may not be the latest one, and our regex was
expecting that.

Relaxing the regex opens for the case where the version on latest is not
the latest, still assuring it is the installed one.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>