docs: update proxies to use http
On howtoguide for configuring proxies, we are using by default
the http prefix for http proxies and https prefix for the https
proxies. We are now changing the default to use http on both
http and https, but adding a note that users might need
to use the https prefix depending on their proxy setup.

Fixes: #2881