Remove docs for format option (#3219)
* docs: remove mention to security-status --format

Since we don't want users to rely on the --format directive anymore,
we are replacing the security-status explanation to instead use
our APIs to get the machine-readable version of running
pro security-status

* docs: remove mention to status --format

We don't want user to rely on pro status --format
anymore. Because of that, we are removing this information
from our docs.