help: don't hide subcommands in usage/help
The metavar information for those commands was being set to empty
string, leading to weird ellipsis in usage and weird blank lines in
help.

There is no reason to hide that for the subcommands, as there will be
usually a sane amount of them. We can revisit if it grows too much but
we should not allow that.

For the main pro --help, that will be completely changed in a subsequent
commit and we are not touching it there now.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>