apt-hook: redirect errors away from users
When warnings or errors appear in the output of apt update or apt
upgrade that are related to our apt hook, it makes pro look bad. This is
true even if the cause of the error is entirely unrelated to pro (see
LP#2055239).

Here we are redirecting all output that should never appear to users to
files that are collected when you run `pro collect-logs`. This way, we
don't lose the messages entirely, and they don't disrupt users.

LP: #2074211