help: expose <command> metavar in the pro usage
This fixes the usage line, justifying the weird ellipsis, and we intend
to use this line when building the help text.

As a side effect, the error message presented when no command is
specified shows that <command> is missing, eliminating the need for the
custom error message pointing to 'help'.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>