apt: always ensure the ESM cache is present
We used to build the base file/folder structure when updating the local
APT cache. However, in some cases, the cache may be accessed before it
is created. In the past an exception was raised and we catched + ignored
it, but from Noble onward this turned to a simple warning on the APT
side.

We fix it by always ensuring the structure is present when trying to
fetch the cache.

Fixes: #3132

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>