detach: fix issue when removing auth file
To verify if we need to remove an entry from
/etc/apt/auth.conf.d/90ubuntu-advantage, we check
if a service url is in that file, if it is, we remove
the line corresponding to the service. However,
we are using the wrong service url to make that check,
since it is missing the /ubuntu suffix. We are now
fixing that issue, allowing detach to remove the appropriate
lines and delete the file if needed