apt-hook: append errors to a log file
The previous method of only keeping the latest potential error output in
a state file was unnecessarily confusing. This new method appends errors
to a log file in /var/log. The already existing logrotate config will
rotate this new file if necessary because of it's wildcard. pro
collect-logs also uses a wildcard and so will also collect this file if
it exists.