api: address transtion packages in vulnerabilities
Better handle how we would capture a transition package
in the vulnerabilities endpoint. By transition, we mean
when a binary package no longer appears under a source package.
In that situation we won't know how to fix the issue, but we
should at least inform the user that the vulnerability affects the
system