tests: don't install old pro package on Noble
On Noble, the ubuntu-advantage-tools package is not installed
by default, that means that once we install the package in our
tests, all of the previous migrations would run. This was causing some
issue on our Pro cloud integration tests. We configure those tests
to disable the auto_attach feature on startup. However, one
of the migrations was overriding our config file momentarily, allowing
our daemon to run and attach the machine. Since we want to avoid
that behavior, we are not installing the ubuntu-advantage-tools
on Noble anymore