help: deduplicate 'pro help' command entries
We use a class variable to make sure each command can register its own
help entry to the main help page. When get_parser is called more than
once, the subparsers are instance bound but the help content gets
duplicated. We now make sure to only add entries that are not present
yet.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>