entitlements: enable USG instead of CIS on Focal
As Focal is the only series to ever have both of the services in a valid
state at different times, there is a particular path to enable USG
instead of CIS when users call `pro enable cis`. A specific message is
shown on the CLI to tell users that it changed.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>