Fix behave output steps
The behave lib was updated recently and now the step matching is more
strict. This mean that if a step has a colon at the end, that colon
should also be reflected on the step implementation as well. We are now
updating all of the output steps related to string matching to reflect
that.