fips: show correct kernel version in the downgrade message on clouds
'linux-fips' was hardcoded as the package to look at. Instead, we look
into whatever the backend tells us is the correct package which will be
installed.

This relies on three facts that must remain true:
- There is only one metapackage for each fips flavor on VMs
- This metapackage is called ubuntu-fips or ubuntu-<something>
- The linux kernel to be installed is called linux-<something>

If something derails, the worse that can happen is, again, a wrong
version in the CLI. So not much to worry about.

Fixes: #3488

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>