apparmor: adjust apparmor profiles to cope with coreutils paths
Due to Rust coreutils in questing, some AppArmor policy is affected
due to the change in path and symlinks.
Instead of using the @{coreutil_dirs} variable, which is available
only on apparmor 5.0.0~alpha1-0ubuntu7, we are using the hardcoded
path. This is because this project is usually backported to older
releases and that will maintain compatibility.

Fixes: https://bugs.launchpad.net/bugs/2123870
Signed-off-by: Georgia Garcia <georgia.garcia@canonical.com>