do-release-upgrade: fix logic to check the APT lock
We need to wait for the lock to be released, but can't hold it while
performing the operations in the hook, because we subprocess to APT,
which needs the lock to run.

Releasing the lock as soon as we get it is good enough, and then we let
APT itself deal with it.

LP: #2129712

Signed-off-by: Renan Rodrigo <rr@ubuntu.com>