chore: remove unused variables from tests
All unused vars away: helper functions, constants, fixtures, parameters,
everything not used in the tests is now gone from the tests.
"But it was there for a reason" reason gone, they gone
"But it reminds us of test gaps" coverage and refactoring too
"But I need to see the past decisions" git log
"But" no

No dead code.

Signed-off-by: Renan Rodrigo <rr@ubuntu.com>