apparmor: fix apparmor denied audit messages
The esm_cache.py script will try to read from
/sys/firmware/devicetree/base/model when /var/lib/ubuntu-advantage/status.json
is not present. On devices without a device tree, the script will simply
print an info message that the file can't be found but on devices with a
device tree, it will generate an apparmor denied audit message as the
apparmor rules don't allow access to this file.

Furthermore, the first run of esm-cache.servce after removal of status.json,
will generate an additional apparmor denied audit message when
systemd-detect-virt is run.

Therefore, add the appropriate rules to the ubuntu_pro_esm_cache and the
ubuntu_pro_esm_cache_systemd_detect_virt apparmor profiles.

Fixes: http://bugs.launchpad.net/bugs/2131292
Signed-off-by: Sebastian Weyer <sebastian.weyer@canonical.com>